marked {dev, test}.env as development

+++ src/db-access/db-access.service.ts
@@ -14,12 +14,13 @@ export class DBAccessService extends PrismaClient {
     });
   }
 
-  // TODO: how to substitute in tests?
   public async clear(): Promise<void> {
-    await this.$transaction([
-      this.profile.deleteMany(),
-      this.user.deleteMany(),
-    ]);
+    if (process.env.DEVELOPMENT) {
+      await this.$transaction([
+        this.profile.deleteMany(),
+        this.user.deleteMany(),
+      ]);
+    }
   }
 
   public static errorCodes = () => {
